Revert "[GPP] Changed GPP to double linked list inspired struct"
This reverts commit 4eb6c9c6e890ccb0dfde96a01ea5f4a9b33a1f92.

+++ inc/types/dictionary.hpp
@@ -27,7 +27,7 @@ typedef struct Dictionary {
   Deletion_Mode deletion_mode;
   Deletion_Handler* deletion_handler;
 private:
-  uint16_t len;             ///< Dictionaries length
+  size_t len;               ///< Dictionaries length
   Dict_entry* first_entry;  ///< First entry - head
   Dict_entry* last_entry;   ///< Last entry - tail
 } Dictionary;
